Complex path improvement and sonar

+++ dmn-core/src/main/java/com/gs/dmn/feel/lib/MixedJavaTimeFEELLib.java
@@ -882,7 +882,7 @@ public class MixedJavaTimeFEELLib extends BaseFEELLib<BigDecimal, LocalDate, Off
         return dateTime.getZone().getId();
     }
     private Duration computeDuration(int secondsOffset) {
-        return DATA_TYPE_FACTORY.newDuration(secondsOffset * 1000);
+        return DATA_TYPE_FACTORY.newDuration((long) secondsOffset * 1000);
     }
 
     //
@@ -1101,7 +1101,7 @@ public class MixedJavaTimeFEELLib extends BaseFEELLib<BigDecimal, LocalDate, Off
             for (int i = 0; i < list.size(); i++) {
                 Object o = list.get(i);
                 if (o == null && match == null || o!= null && o.equals(match)) {
-                    result.add(BigDecimal.valueOf(i + 1));
+                    result.add(BigDecimal.valueOf((long) i + 1));
                 }
             }
         }
